depi-impl/depi/monitors/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "setuptools-grpc"]
build-backend = "setuptools.build_meta"

[project]
name = "depi-monitors"
version="0.7.7"
authors = [
    {name = "Mark Wutka", email="mark.wutka@vanderbilt.edu"},
    {name = "Patrik Meijer", email="patrik.meijer@vanderbilt.edu"}
]
requires-python = ">=3.10"
dependencies=[
    "build==1.0.3",
    "Flask==2.3.2",
    "gitdb==4.0.10",
    "GitPython==3.1.31",
    "grpcio==1.64.0",
    "grpcio-tools==1.64.0",
    "protobuf==5.27.0",
    "pylint==2.17.5",
    "textX==3.1.1",
    "urllib3==2.0.4",
    "virtualenv==20.23.0",
    "waitress==2.1.2",
]

[project.scripts]
depi-git-monitor="depi_monitors.git_monitor:run"
depi-git-adaptor="depi_monitors.git_adaptor:run"
depi-git-verifier="depi_monitors.git_verifier:run"
depi-gsn-adaptor="depi_monitors.gsn_adaptor:run"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
depi_monitors = ["../depi_pb2.py", "../depi_pb2.pyi", "../depi_pb2_grpc.py",
    "*.tx"]
gsn_monitor = [ "*.tx"]
depi_pb2 = []
depi_pb2_grpc = []
